+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [2.28.0] - 2021-09-
+## [2.28.0] - 2021-09-20
 ### Changed
 - Choosing first user account by default if there are multiple, not supporting multiple accounts for now
 
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactored Express server using HTML/CSS for better UX
 - Improved UX for asking if repo should be public or private
 - Added Loading... view for left panel
-- Menu options change on the fly with 100% coverage
+- Menu options change on the fly
 
 ## [2.27.0] - 2021-09-14
 ### Fixed